update content App, add HLS code

+++ content/api/signals.py
@@ -1,9 +1,12 @@
 import os
 import django_rq
+import shutil
 from django.dispatch import receiver
 from django.db.models.signals import post_save, post_delete
+from django.conf import settings
 from content.models import Video
-from content.tasks import convert_to_resolution
+from content.tasks import convert_to_hls
+
 
 @receiver (post_save, sender=Video)
 def video_post_save(sender, instance, created, **kwargs):
@@ -11,14 +14,18 @@ def video_post_save(sender, instance, created, **kwargs):
         source = instance.video_file.path
         if os.path.exists(source):
             queue  = django_rq.get_queue('default', autocommit=True)
-            queue.enqueue(convert_to_resolution, source)
+            queue.enqueue(convert_to_hls, source, instance.id)
 
-@receiver(post_delete, sender=Video)
-def auto_delete_file_on_delete(sender, instance, **kwargs):
-    if instance.video_file:
-        if os.path.isfile(instance.video_file.path):
-            os.remove(instance.video_file.path)
-    
-    if instance.thumbnail:
-        if os.path.isfile(instance.thumbnail.path):
-            os.remove(instance.thumbnail.path)
+
+@receiver(post_delete, sender=Video)       
+def auto_delete_files_on_video_delete(sender, instance, **kwargs):
+   
+    if instance.video_file and os.path.isfile(instance.video_file.path):
+        os.remove(instance.video_file.path)
+
+    hls_dir = os.path.join(instance.video_file.storage.location, 'videos', 'hls', str(instance.id))
+    if os.path.isdir(hls_dir):
+        shutil.rmtree(hls_dir)
+
+    if instance.thumbnail_url and os.path.isfile(instance.thumbnail_url.path):
+        os.remove(instance.thumbnail_url.path)
